requirements.txt
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
schedule==1.2.1
pandas==2.1.3
cryptography==41.0.5
tqdm==4.66.1
fake-useragent==1.3.0
